ai-agent/requirements.txt
x402
eth-account
httpx
python-dotenv

